Added Privacy Statement
Added the Privacy Statement from the UCI Directory. Privacy Notice:
This campus directory has been compiled for the use and convenience of
the faculty, staff and students of the University of California, Irvine
and others dealing with UCI. It is the property of the Regents of the
University of California. In accordance with the California Information
Practices Act, neither this directory nor the information contained
herein may be used, rented, distributed, or sold for commercial
purposes.